00_Data_Ingestion/requirements.txt
pandas
boto3
kafka-python
python-dotenv
numpy
requests
atproto